Deprecated Util, need to get updated math lib for Mod
